move async code to main class
